+++ src/dynamixel_control_hw.cpp
@@ -41,6 +41,12 @@
 // for dynamixel::OperatingMode
 #include <dynamixel/servos/base_servo.hpp>
 
+// to parse parameters
+#include <XmlRpcValue.h>
+#include <XmlRpcException.h>
+
+dynamixel::OperatingMode get_mode(const std::string& mode_string, std::string name);
+
 int main(int argc, char** argv)
 {
     ros::init(argc, argv, "dynamixel_control_hw");
@@ -65,68 +71,84 @@ int main(int argc, char** argv)
     got_all_params &= nhParams.getParam("serial_interface", usb_serial_interface);
     int baudrate; // in bauds
     got_all_params &= nhParams.getParam("baudrate", baudrate);
-    double dynamixel_timeout; // in seconds
-    got_all_params &= nhParams.getParam("dynamixel_timeout", dynamixel_timeout);
-    double dynamixel_scanning; // in seconds
-    bool dyn_scan = nhParams.getParam("dynamixel_scanning", dynamixel_scanning);
+    double read_timeout; // in seconds
+    got_all_params &= nhParams.getParam("read_timeout", read_timeout);
+    double scan_timeout; // in seconds
+    bool dyn_scan = nhParams.getParam("scan_timeout", scan_timeout);
     if (!dyn_scan) {
         ROS_WARN_STREAM("Dynamixel scanning timeout parameter was not found. "
             << "Setting to default: 0.05s.");
-        dynamixel_scanning = 0.05;
+        scan_timeout = 0.05;
     }
+    std::string default_command_interface;
+    bool has_default_command_interface = nhParams.getParam(
+        "default_command_interface", default_command_interface);
 
-    // Retrieve the map from joint name to hardware-related ID
-    // It has to be inverted, putting the ID as a key, for later use
+    // Map from joint name to hardware-related ID
     std::unordered_map<Protocol::id_t, std::string> dynamixel_map;
-    std::map<std::string, int> map_param; // temporary map, from parameter server
-    got_all_params &= nhParams.getParam("hardware_mapping", map_param);
-    std::map<std::string, int>::iterator map_param_i;
-    for (map_param_i = map_param.begin(); map_param_i != map_param.end(); map_param_i++) {
-        dynamixel_map[(Protocol::id_t)map_param_i->second] = map_param_i->first;
-    }
-
-    // Retrieve the map for command interface (ID: velocity/position)
+    // Map for command interface (ID: velocity/position)
     std::unordered_map<Protocol::id_t, dynamixel::OperatingMode> dynamixel_c_mode_map;
-    std::map<std::string, std::string> c_mode_param; // temporary map, from parameter server
-    got_all_params &= nhParams.getParam("command_interface", c_mode_param);
-    std::map<std::string, std::string>::iterator c_mode_param_i;
-    for (c_mode_param_i = c_mode_param.begin(); c_mode_param_i != c_mode_param.end(); c_mode_param_i++) {
-        long long int k;
-        std::istringstream(c_mode_param_i->first) >> k;
-
-        dynamixel::OperatingMode mode;
-        if (c_mode_param_i->second == "velocity")
-            mode = dynamixel::OperatingMode::wheel;
-        else if (c_mode_param_i->second == "position")
-            mode = dynamixel::OperatingMode::joint;
-        else {
-            ROS_FATAL_STREAM("The command mode " << c_mode_param_i->second
-                                                 << " is not available (actuator "
-                                                 << k << ")");
-            return 1;
-        }
-
-        dynamixel_c_mode_map[k] = mode;
-    }
-
-    // Retrieve the map for max speed (ID: max speed)
-    std::unordered_map<Protocol::id_t, double> dynamixel_max_speed_map;
-    std::map<std::string, double> max_speed_param; // temporary map, from parameter server
-    nhParams.getParam("max_speed", max_speed_param);
-    std::map<std::string, double>::iterator max_speed_param_i;
-    for (max_speed_param_i = max_speed_param.begin(); max_speed_param_i != max_speed_param.end(); max_speed_param_i++) {
-        Protocol::id_t k = std::stoll(max_speed_param_i->first);
-        dynamixel_max_speed_map[k] = max_speed_param_i->second;
-    }
-
-    // Retrieve the map with angle corrections (ID: correction in radians)
+    // Map with angle corrections (ID: correction in radians)
     std::unordered_map<Protocol::id_t, double> dynamixel_corrections;
-    std::map<std::string, double> map_corrections; // temporary map, from parameter server
-    nhParams.getParam("hardware_corrections", map_corrections);
-    std::map<std::string, double>::iterator map_cor_i;
-    for (map_cor_i = map_corrections.begin(); map_cor_i != map_corrections.end(); map_cor_i++) {
-        Protocol::id_t k = std::stoll(map_cor_i->first);
-        dynamixel_corrections[k] = map_cor_i->second;
+    // Map for max speed (ID: max speed)
+    std::unordered_map<Protocol::id_t, double> dynamixel_max_speed;
+
+    // Retrieve the srevo parameter and fill maps above with its content.
+    std::unordered_map<Protocol::id_t, std::string> servos_map;
+    XmlRpc::XmlRpcValue servos_param; // temporary map, from parameter server
+    got_all_params &= nhParams.getParam("servos", servos_param);
+    ROS_ASSERT(servos_param.getType() == XmlRpc::XmlRpcValue::TypeStruct);
+    try {
+        for (XmlRpc::XmlRpcValue::ValueStruct::const_iterator it = servos_param.begin(); it != servos_param.end(); ++it) {
+            ROS_DEBUG_STREAM("Found servo: " << (std::string)(it->first));
+
+            Protocol::id_t id;
+            if (it->second.hasMember("id")) {
+                id = static_cast<int>(servos_param[it->first]["id"]);
+                ROS_DEBUG_STREAM("\tid: " << servos_param[it->first]["id"]);
+                dynamixel_map[id] = it->first;
+            }
+            else {
+                ROS_ERROR_STREAM("Actuator " << it->first
+                                             << " has no associated servo ID.");
+            }
+
+            if (it->second.hasMember("offset")) {
+                ROS_DEBUG_STREAM("\toffset: "
+                    << servos_param[it->first]["offset"]);
+                dynamixel_corrections[id]
+                    = static_cast<double>(servos_param[it->first]["offset"]);
+            }
+            if (it->second.hasMember("max_speed")) {
+                ROS_DEBUG_STREAM("\tmax_speed: "
+                    << servos_param[it->first]["max_speed"]);
+                dynamixel_max_speed[id]
+                    = static_cast<double>(servos_param[it->first]["max_speed"]);
+            }
+
+            if (it->second.hasMember("command_interface")) {
+                std::string mode_string
+                    = static_cast<std::string>(servos_param[it->first]["command_interface"]);
+                ROS_DEBUG_STREAM("\tcommand_interface: " << mode_string);
+
+                dynamixel_c_mode_map[id] = get_mode(mode_string, it->first);
+            }
+            else if (has_default_command_interface) {
+                ROS_DEBUG_STREAM("\tcommand_interface: "
+                    << default_command_interface << " (default)");
+                dynamixel_c_mode_map[id]
+                    = get_mode(default_command_interface, it->first);
+            }
+            else {
+                ROS_ERROR_STREAM("A command interface (speed or position) "
+                    << "should be declared for the actuator " << it->first
+                    << " or a default one should be defined with the parameter "
+                    << "'default_command_interface'.");
+            }
+        }
+    }
+    catch (XmlRpc::XmlRpcException& e) {
+        ROS_FATAL_STREAM("Exception raised when parsing the configuration: " << e.getMessage());
     }
 
     if (!got_all_params) {
@@ -134,9 +156,8 @@ int main(int argc, char** argv)
                                     "were not set:\n"
             + sub_namespace + "/serial_interface\n"
             + sub_namespace + "/baudrate\n"
-            + sub_namespace + "/dynamixel_timeout\n"
-            + sub_namespace + "/hardware_mapping\n"
-            + sub_namespace + "/command_interface";
+            + sub_namespace + "/read_timeout\n"
+            + sub_namespace + "/servos";
         ROS_FATAL_STREAM(error_message);
         return 1;
     }
@@ -146,6 +167,7 @@ int main(int argc, char** argv)
 
     // We run the ROS loop in a separate thread as external calls, such
     // as service callbacks loading controllers, can block the (main) control loop
+
     ros::AsyncSpinner spinner(2);
     spinner.start();
 
@@ -154,11 +176,11 @@ int main(int argc, char** argv)
         dynamixel_hw_interface = std::make_shared<dynamixel::DynamixelHardwareInterface<Protocol>>(
             usb_serial_interface,
             baudrate,
-            dynamixel_timeout,
-            dynamixel_scanning,
+            read_timeout,
+            scan_timeout,
             dynamixel_map,
             dynamixel_c_mode_map,
-            dynamixel_max_speed_map,
+            dynamixel_max_speed,
             dynamixel_corrections);
     dynamixel_hw_interface->init();
 
@@ -169,4 +191,27 @@ int main(int argc, char** argv)
     ros::waitForShutdown();
 
     return 0;
+}
+
+/** Convert a string to an operating mode for a Dynamixel servo
+
+    @param mode_string name of the operating mode (either velocity or position)
+    @param nam name of the joint
+    @return dynamixel::OperatingMode::unknown if mode_string is not recognized
+**/
+dynamixel::OperatingMode get_mode(const std::string& mode_string, std::string name)
+{
+    dynamixel::OperatingMode mode;
+    if ("velocity" == mode_string)
+        mode = dynamixel::OperatingMode::wheel;
+    else if ("position" == mode_string)
+        mode = dynamixel::OperatingMode::joint;
+    else {
+        mode = dynamixel::OperatingMode::unknown;
+        ROS_ERROR_STREAM("The command mode " << mode_string
+                                             << " is not available (actuator "
+                                             << name << ")");
+    }
+
+    return mode;
 }